Fix front camera open fail due to TOF driver running always
For Hannah which TOF supported , if the camera daemon is died when
rear camera is running, the TOF driver will not close on time,
which will lead to next use of close rear camera can not be done at
stats module. So change the TOF driver to close status when daemen died.

Change-Id: I41669b3ce1ebaaad6566c9b30e4c8cd7cbe884c3
Signed-off-by: libm1 <libm1@lenovo.com>
Reviewed-on: https://gerrit.mot.com/1114638
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Dawei Wang <wangdw10@motorola.com>
Reviewed-by: Biming Li <libm1@motorola.com>
Reviewed-by: Chenxi Tan <tancx1@motorola.com>
Reviewed-by: Xiangpo Zhao <zhaoxp3@motorola.com>
Submit-Approved: Jira Key