msm: mdss: Set backlight on unblank after 2nd frame is committed
Don't enable backlight after unblank until after 2nd frame is committed in
order to ensure that display contents are defined and the display is fully
powered on and ready to render the frame contents. If the 2nd frame is not
committed within 15 milliseconds, then the backlight will be enabled
regardless.

This should resolve the framebuffer garbage seen during system resume for
some users.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: Raphiel Rollerscaperers <raphielscape@outlook.com>