power: qpnp-smbcharger: Hide charge counter on selected targets
Some devices have a bug where querying the charge counter during deep
sleep takes an excessively long amount of time. BatteryService looks
up this prop during deep sleep, resulting in a very slow wakeup on these
devices.

land has this bug while santoni does not. Add a filter to hide the
charge counter on affected targets, so BatteryService does not attempt
to read the prop. It is unknown as to whether this is a firmware or a
hardware issue.

Signed-off-by: Albert I <krascgq@outlook.co.id>
Signed-off-by: utsavbalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: utsavisgreat <utsavbalar1231@gmail.com>
Signed-off-by: UtsavisGreat <utsavbalar1231@gmail.com>